src/app/shared/components/checkout/cost-center-selection/cost-center-selection.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { TranslateModule } from '@ngx-translate/core';

import { FormlyTestingComponentsModule } from 'ish-shared/formly/dev/testing/formly-testing-components.module';

import { CostCenterSelectionComponent } from './cost-center-selection.component';

describe('Cost Center Selection Component', () => {
  let component: CostCenterSelectionComponent;
  let fixture: ComponentFixture<CostCenterSelectionComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [CostCenterSelectionComponent],
      imports: [FormlyTestingComponentsModule, TranslateModule.forRoot()],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(CostCenterSelectionComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
